integrate elevation function to the data pipeline. Coloring with texts need to be fixed.
