+++ cartograph/graph/Edge.py
@@ -1,0 +1,11 @@
+class Edge:
+    """
+    An edge which records two centers and two adjacent corners
+    """
+    def __init__(self, id, center1, center2, vertex1, vertex2, is_border):
+        self.id = id
+        self.d0 = center1
+        self.d1 = center2
+        self.v0 = vertex1
+        self.v1 = vertex2
+        self.is_border = is_border
